requirements.txt
opencv-python>=4.5.0
pillow>=8.0.0
pytesseract>=0.3.8
numpy>=1.19.0
PyQt5>=5.15.0
PyQt5-sip>=12.8.0
